move crossScalaVersions to commonSettings; move verified contracts test to scala 2.12 only source folder; add implicit conversions from scala lambdas to java Function and Consumer;
