+++ appkit/src/test/scala/org/ergoplatform/appkit/examples/RunMockedScala.scala
@@ -2,7 +2,7 @@ package org.ergoplatform.appkit.examples
 
 import java.io.File
 
-import org.ergoplatform.appkit.{AppkitTesting, JavaHelpers}
+import org.ergoplatform.appkit.{AppkitTesting, BlockchainContext, JavaHelpers}
 import org.ergoplatform.appkit.examples.util.FileMockedErgoClient
 import scalan.util.FileUtil
 import JavaHelpers._
@@ -17,7 +17,7 @@ object RunMockedScala extends App with AppkitTesting {
 
   val res = new FileMockedErgoClient(
     nodeResponses,
-    java.util.Arrays.asList()).execute { ctx =>
+    java.util.Arrays.asList()).execute { ctx: BlockchainContext =>
     val r = new ExampleScenarios(ctx)
     val res = r.aggregateUtxoBoxes("abc", addrStr, 10, "83b94f2df7e97586a9fe8fe43fa84d252aa74ecee5fe0871f85a45663927cd9a")
     res
